feat: add item from product page functionality
